+++ okhttp/src/main/kotlin/okhttp3/internal/socket/OkioServerSocketFactory.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2024 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package okhttp3.internal.socket
+
+import javax.net.ServerSocketFactory
+
+interface OkioServerSocketFactory {
+  fun newServerSocket(): OkioServerSocket
+}
+
+class RealOkioServerSocketFactory(
+  private val delegate: ServerSocketFactory = ServerSocketFactory.getDefault(),
+) : OkioServerSocketFactory {
+  override fun newServerSocket(): OkioServerSocket {
+    val serverSocket = delegate.createServerSocket()
+    return RealOkioServerSocket(serverSocket)
+  }
+}
